fix: クエリとお気に入りの実体となる構造体をbotパッケージへ移動
